requirements.txt
praw
python-dotenv
requests>=2.25.0 